Refine Entity
