+++ 02_CODE_GameLib/Interfaces/Entity/IPlayer.cs
@@ -1,14 +1,17 @@
 using System.Collections.Generic;
+using CODE_GameLib.Enums;
 using CODE_GameLib.Interfaces.Items.Wearable;
 
 namespace CODE_GameLib.Interfaces.Entity
 {
     public interface IPlayer : IEntity
     {
+        public List<IWearable> Inventory { get; }
+
+        public List<Cheat> Cheats { get; }
+        
         public bool Won { get; }
-
-        public IEnumerable<IWearable> Inventory { get; }
-
+        
         public bool AddToInventory(IWearable wearable);
