+++ 02_CODE_GameLib/Observers/PlayerObserver.cs
@@ -4,7 +4,7 @@ using CODE_GameLib.Interfaces.Entity;
 
 namespace CODE_GameLib.Observers
 {
-    public class PlayerObserver : IObserver<IPlayer>, IObserver<IEntity>
+    public class PlayerObserver : IObserver<IEntity>
     {
 
         private readonly IGame _game;
@@ -28,11 +28,6 @@ namespace CODE_GameLib.Observers
         {
             if (entity.Died || (entity is IPlayer player && player.Won))
                 _game.Destroy();
-        }
-
-        public void OnNext(IPlayer player)
-        {
-            
         }
     }
 }
