+++ 02_CODE_GameLib/Room.cs
@@ -8,6 +8,8 @@ namespace CODE_GameLib
     {
         public int Width { get; }
         public int Height { get; }
+        public int CenterX => (Width + 1) / 2 - 1;
+        public int CenterY => (Height + 1) / 2 - 1;
         public List<IItem> Items { get; }
         public IEnumerable<IConnection> Connections { get; }
